+++ lib/index.ts
@@ -1,6 +1,6 @@
 export * from "./schematic/CircuitJsonToKicadSchConverter"
 export * from "./pcb/CircuitJsonToKicadPcbConverter"
 export * from "./project/CircuitJsonToKicadProConverter"
-export * from "./project/resolveAndLoadKicad3dModelFiles"
+export * from "./utils/resolveAndLoadKicad3dModelFiles"
 export * from "./kicad-library/CircuitJsonToKicadLibraryConverter"
 export * from "./kicad-library/KicadLibraryConverter"
