+++ lib/utils/resolveAndLoadKicad3dModelFiles.ts
